fix: re-fix race condition on global config
The last commit didn't seem to fully fix the issue, which in part seems
to be caused by watching array values in globalConfig. After updating
sourceFieldIds, it would initially update, and then a second later a
message that wasn't triggered by any of the updates here would fire in
the console under "message received" that replaced the value with an
empty array. Switching to string values and handling updates outside of
useEffect seems to address it after more testing.